AOS data extraction refactor to bring in line with DN data extraction
